Separate decorator classes
