+++ nameable.rb
@@ -3,26 +3,3 @@ class Nameable
     raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
   end
 end
-
-class Decorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.length > 10 ? @nameable.correct_name[0..9] : @nameable.correct_name
-  end
-end
